requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
redis
flask_cors
flask_login
datetime
flask-jwt-extended